requirements.txt
pytest==7.0.1
selenium==4.1.3
python-dotenv==0.19.2
allure-pytest==2.9.45
Faker==13.3.1
pytest-xdist==2.5.0
requests==2.27.1